LPS-122456 - Remove createContextualFragment in senna for now
